Speed-up
The Last.fm gettopartists call returns URLs of images, so there's no need to make another 1000 or so calls to Echo Nest or Last.fm to ask for the URL again. Makes it about five times quicker.

pylast didn't have images in the gettopartists, so just get the JSON directly and parse it ourselves, it's the only API call we need. So removed pylast and pyen.

Also removed user playcount weightings.